pyproject.toml
[tool.poetry]
name = "BorAI"
version = "0.3.0"
description = "Discord AI using the latest LLM tech"
authors = ["Hegedüs Armand Alex <hearlex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
py-cord = "2.5.0"
python-dotenv = "^1.0.0"
openai = "^1.23.2"
termcolor = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
